+++ src/main/kotlin/io/snyk/plugin/ui/settings/ScanTypesPanel.kt
@@ -142,6 +142,7 @@ class ScanTypesPanel(
             }
         }
     }.apply {
+        name = "scanTypesPanel"
         border = JBUI.Borders.empty(2)
     }